+++ src/components/Education/components/index.js
@@ -1,2 +1,0 @@
-export { default as DegreeDetails } from './degreeDetails';
-export { default as EducationCard } from './educationCard';