+++ src/components/Experience/components/index.js
@@ -1,2 +1,0 @@
-export { default as ExperienceCard } from './experienceCard';
-export { default as PositionDetails } from './positionDetails';