+++ src/components/Experience/experienceStyles.js
@@ -1,35 +1,7 @@
 import styled from 'styled-components';
 import { colors } from '../../utils/styleVariables';
 
-export const SectionContainer = styled.div`
-  padding: 7vh;
-
-  h2 {
-    text-align: center;
-    margin: 3vh;
-  }
-`;
-
-export const ElementContainer = styled.div`
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-`;
-
-export const CardContainer = styled.div`
-  box-shadow: 0px 5px 25px 0px rgba(0, 0, 0, 0.2);
-  display: flex;
-  flex-direction: row;
-  margin: 2vh;
-  width: 75vw;
-
-  :hover {
-    transform: scale(1.01, 1.01);
-    transition: 0.5s;
-  }
-`;
-
+// eslint-disable-next-line import/prefer-default-export
 export const DetailsContainer = styled.div`
   padding: 4vh;