+++ src/components/index.js
@@ -1,4 +1,5 @@
 export { default as Articles } from './Articles';
+export { default as Awards } from './Awards';
 export { default as Education } from './Education';
 export { default as Experience } from './Experience';
 export { default as Introduction } from './Introduction';